rename plant module NetworkBettina to BettinaNetwork
