+++ Benchmarks/ModuleBenchmarks/PlantModules/BettinaNetwork/Belowground/NetworkFixedSalinity/NetworkFixedSalinity_CI.xml
@@ -5,12 +5,15 @@
             <type> SimpleTest </type>
         </aboveground>
         <belowground>
-            <type> SimpleNetwork </type>
+            <type> NetworkFixedSalinity </type>
             <f_radius> 0.25 </f_radius>
+            <min_x>0</min_x>
+            <max_x>22</max_x>
+            <salinity>0.025 0.035</salinity>
         </belowground>
     </resources>
     <plant_dynamics>
-        <type> NetworkBettina </type>
+        <type> BettinaNetwork </type>
         <mortality>NoGrowth</mortality>
         <f_growth> 0.25 </f_growth>
         <variant> V2_adapted </variant>
@@ -22,7 +25,7 @@
             <distribution>
                 <type> GroupFromFile </type>
                 <n_recruitment_per_step> 0 </n_recruitment_per_step>
-                <filename> Benchmarks/ModuleBenchmarks/PlantModules/NetworkBettina/bg_initial_population.csv </filename>
+                <filename> Benchmarks/ModuleBenchmarks/PlantModules/BettinaNetwork/bg_initial_population.csv </filename>
             </distribution>
         </group>
     </population>
@@ -48,6 +51,7 @@
         <growth_output> growth </growth_output>
         <growth_output> ag_resources </growth_output>
         <growth_output> bg_resources </growth_output>
+        <network_output> psi_osmo </network_output>
         <network_output> rgf </network_output>
         <network_output> partner </network_output>
         <network_output> potential_partner </network_output>
