+++ Benchmarks/ModuleBenchmarks/PlantModules/BettinaNetwork/Belowground/NetworkOGSLargeScale3D/ReferenceFiles/NetworkOGSLargeScale3D/Population_t_0002000000.0.csv
