+++ Benchmarks/ModuleBenchmarks/PlantModules/BettinaNetwork/Belowground/NetworkOGSLargeScale3D/python_script.py
