+++ ProjectLib/Project.py
@@ -97,8 +97,8 @@ class MangaProject:
             from PlantModelLib.Bettina import Bettina as createGD
         elif case == "SimpleKiwi":
             from PlantModelLib.SimpleKiwi import SimpleKiwi as createGD
-        elif case == "NetworkBettina":
-            from PlantModelLib.NetworkBettina import NetworkBettina as createGD
+        elif case == "BettinaNetwork":
+            from PlantModelLib.BettinaNetwork import BettinaNetwork as createGD
         else:
             raise KeyError("Required plant dynamic concept not implemented.")
         self.plant_dynamic_concept = createGD(arg)
